format with `ruff`

+++ tests/test_evaluation.py
@@ -6,10 +6,12 @@ import pandas as pd
 from utilsforecast.losses import *
 from utilsforecast.data import generate_series
 from utilsforecast.evaluation import evaluate
+import utilsforecast.processing as ufp
+import polars as pl
 
 series = generate_series(10, n_models=2, level=[80, 95])
-series['unique_id'] = series['unique_id'].astype('int')
-models = ['model0', 'model1']
+series["unique_id"] = series["unique_id"].astype("int")
+models = ["model0", "model1"]
 metrics = [
     mae,
     mse,
@@ -30,27 +32,27 @@ evaluation = evaluate(
     train_df=series,
     level=[80, 95],
 )
-summary = evaluation.drop(columns='unique_id').groupby('metric').mean().reset_index()
+summary = evaluation.drop(columns="unique_id").groupby("metric").mean().reset_index()
 summary
 import polars.testing
 
-series_pl = generate_series(10, n_models=2, level=[80, 95], engine='polars')
+series_pl = generate_series(10, n_models=2, level=[80, 95], engine="polars")
 pl_evaluation = evaluate(
     series_pl,
     metrics=metrics,
     train_df=series_pl,
     level=[80, 95],
-).drop('unique_id')
-pl_summary = ufp.group_by(pl_evaluation, 'metric').mean()
+).drop("unique_id")
+pl_summary = ufp.group_by(pl_evaluation, "metric").mean()
 pd.testing.assert_frame_equal(
-    summary.sort_values('metric'),
-    pl_summary.sort('metric').to_pandas(),
+    summary.sort_values("metric"),
+    pl_summary.sort("metric").to_pandas(),
 )
 pl.testing.assert_frame_equal(
     evaluate(
-        series_pl, metrics=metrics, train_df=series_pl, level=[80, 95], agg_fn='mean'
-    ).sort('metric'),
-    pl_summary.sort('metric'),
+        series_pl, metrics=metrics, train_df=series_pl, level=[80, 95], agg_fn="mean"
+    ).sort("metric"),
+    pl_summary.sort("metric"),
 )
 from datasetsforecast.evaluation import accuracy as ds_evaluate
 import datasetsforecast.losses as ds_losses
@@ -61,7 +63,7 @@ def daily_mase(y, y_hat, y_train):
 
 
 level = [80, 95]
-for agg_fn in [None, 'mean']:
+for agg_fn in [None, "mean"]:
     uf_res = evaluate(
         series,
         metrics=metrics,
@@ -70,7 +72,7 @@ for agg_fn in [None, 'mean']:
         level=level,
         agg_fn=agg_fn,
     )
-    agg_by = None if agg_fn == 'mean' else ['unique_id']
+    agg_by = None if agg_fn == "mean" else ["unique_id"]
     ds_res = ds_evaluate(
         series,
         metrics=[
@@ -90,23 +92,23 @@ for agg_fn in [None, 'mean']:
         Y_df=series,
         agg_by=agg_by,
     )
-    ds_res['metric'] = ds_res['metric'].str.replace('-', '_')
-    ds_res['metric'] = ds_res['metric'].str.replace('q_', 'q')
-    ds_res['metric'] = ds_res['metric'].str.replace('lv_', 'level')
-    ds_res['metric'] = ds_res['metric'].str.replace('daily_mase', 'mase')
+    ds_res["metric"] = ds_res["metric"].str.replace("-", "_")
+    ds_res["metric"] = ds_res["metric"].str.replace("q_", "q")
+    ds_res["metric"] = ds_res["metric"].str.replace("lv_", "level")
+    ds_res["metric"] = ds_res["metric"].str.replace("daily_mase", "mase")
     # utils doesn't multiply pct metrics by 100
-    ds_res.loc[ds_res['metric'].str.startswith('coverage'), ['model0', 'model1']] /= 100
-    ds_res.loc[ds_res['metric'].eq('mape'), ['model0', 'model1']] /= 100
+    ds_res.loc[ds_res["metric"].str.startswith("coverage"), ["model0", "model1"]] /= 100
+    ds_res.loc[ds_res["metric"].eq("mape"), ["model0", "model1"]] /= 100
     # we report smape between 0 and 1 instead of 0-200
-    ds_res.loc[ds_res['metric'].eq('smape'), ['model0', 'model1']] /= 200
+    ds_res.loc[ds_res["metric"].eq("smape"), ["model0", "model1"]] /= 200
 
     ds_res = ds_res[uf_res.columns]
     if agg_fn is None:
-        ds_res = ds_res.sort_values(['unique_id', 'metric'])
-        uf_res = uf_res.sort_values(['unique_id', 'metric'])
+        ds_res = ds_res.sort_values(["unique_id", "metric"])
+        uf_res = uf_res.sort_values(["unique_id", "metric"])
     else:
-        ds_res = ds_res.sort_values('metric')
-        uf_res = uf_res.sort_values('metric')
+        ds_res = ds_res.sort_values("metric")
+        uf_res = uf_res.sort_values("metric")
 
     pd.testing.assert_frame_equal(
         uf_res.reset_index(drop=True),
@@ -121,7 +123,7 @@ from pyspark.sql import SparkSession
 
 if sys.version_info >= (3, 9):
     spark = SparkSession.builder.getOrCreate()
-    spark.sparkContext.setLogLevel('FATAL')
+    spark.sparkContext.setLogLevel("FATAL")
     dask_df = dd.from_pandas(series, npartitions=2)
     spark_df = spark.createDataFrame(series).repartition(2)
     for distributed_df, use_train in product([dask_df, spark_df], [True, False]):
@@ -142,7 +144,7 @@ if sys.version_info >= (3, 9):
             )
         ).as_pandas()
         pd.testing.assert_frame_equal(
-            local_res.sort_values(['unique_id', 'metric']).reset_index(drop=True),
-            distr_res.sort_values(['unique_id', 'metric']).reset_index(drop=True),
+            local_res.sort_values(["unique_id", "metric"]).reset_index(drop=True),
+            distr_res.sort_values(["unique_id", "metric"]).reset_index(drop=True),
             check_dtype=False,
         )
